refactor: split decoder builder into separate file
Move the DecoderBuilder implementation into its own module file to improve
code organization and maintainability. The builder pattern is now contained
in decoder/builder.rs while keeping the core Decoder types and traits in
decoder/mod.rs.

This change:
- Extracts builder-specific code to builder.rs
- Updates documentation to reflect the split
- Maintains existing functionality and API
